Refactor: Move all interfaces and enums to "types" directory
Renamed and reorganized files under `interfaces` and `enums` into a unified `types` directory for clearer project structure. This improves maintainability and aligns naming conventions across the codebase. Minor formatting adjustments were also made in several files.
