--- src/app/core/interfaces/JWTInformation.ts
@@ -1,14 +1,0 @@
-export interface JWTInformation {
-	rawToken: string;
-	groups: string[];
-	name: string;
-	claims: JWTClaims[];
-	issuer: string;
-	issuedAtTime: number;
-	expirationTime: number;
-}
-
-interface JWTClaims {
-	name: string;
-	value: string;
-}
