+++ src/app/core/types/dto/announcement/announcement.ts
