+++ src/app/core/types/dto/forms/AddTeacher.ts
@@ -1,5 +1,3 @@
-import {FormControl, Validators} from "@angular/forms";
-
 export interface AddTeacherDTO {
   firstName: string;
   lastName: string;
