+++ src/app/core/types/dto/guardian/Guardian.ts
@@ -1,4 +1,4 @@
-import { Student } from "../student/Student";
+import {Student} from "../student/Student";
 
 export interface Guardian {
   id: number;
