+++ src/app/core/types/dto/strand/Strand.ts
