+++ src/app/core/types/dto/student/Student.ts
@@ -1,7 +1,7 @@
-import { Sex } from '../../../enums/Sex';
-import { GradeLevel } from '../grade-level/GradeLevel';
-import { Strand } from '../strand/Strand';
-import { Teacher } from '../teacher/Teacher';
+import {Sex} from '../../enums/Sex';
+import {GradeLevel} from '../grade-level/GradeLevel';
+import {Strand} from '../strand/Strand';
+import {Teacher} from '../teacher/Teacher';
 
 // Define the missing interfaces
 interface StudentClassroom {
