+++ src/app/core/types/enums/AttendanceForeignEntity.ts
