+++ src/app/core/types/enums/TimeRange.ts
