+++ src/app/core/types/enums/TimeStack.ts
