+++ src/app/core/types/other/EventItem.ts
