+++ src/app/core/types/other/LineChartDTO.ts
