+++ src/app/core/types/other/MessageDTO.ts
